Refactor to move some sync types to new domain folder
