+++ src/components/sync-wizard/metadata/MetadataIncludeExcludeStep.tsx
@@ -7,7 +7,7 @@ import i18n from "../../../locales";
 import { D2Model } from "../../../models/dhis/default";
 import { d2ModelFactory } from "../../../models/dhis/factory";
 import { D2, ModelDefinition } from "../../../types/d2";
-import { MetadataPackage } from "../../../types/synchronization";
+import { MetadataPackage } from "../../../domain/synchronization/Entities";
 import { getMetadata } from "../../../utils/synchronization";
 import Dropdown, { DropdownOption } from "../../dropdown/Dropdown";
 import { Toggle } from "../../toggle/Toggle";
