+++ src/data/synchronization/mappers/PackageTransformations.ts
@@ -1,5 +1,5 @@
 import { PackageTransformationStrategy } from "./D2VersionPackageMapper";
-import { MetadataPackage, EventsPackage, AggregatedPackage } from "../../../types/synchronization";
+import { MetadataPackage, EventsPackage, AggregatedPackage } from "../../../domain/synchronization/Entities";
 
 export const metadataTransformations: PackageTransformationStrategy<MetadataPackage>[] = [];
 export const eventsTransformations: PackageTransformationStrategy<EventsPackage>[] = [];
